Added the basic PetOwner npc behavior graph
To do this I had to implement a few new systems. I added the GetZonePoint gate state to select a random point in a zone that I use to walk randomly. Then we have the BoolSelect gate state which chooses an outcome based on a bool variable. SequentialSayState loops through a list of speech bubbles, one each time it is activated. To get this to work I had to add the ability to save arbitrary data to the npcContext. States access this to persist data across activations since each new state is a new object even if the state previously ran.