Added scanner interactable that gives the scanned role if you are not infected.
if you are infected it gives level 10 suspicion foreverish. It also exposes an event that will be used to trigger a bool that controls alarm activation. Perhaps an entire AlarmManager so that it can track sources of the alarm and correctly disable them when the npc that created the alarm dies or suspicion goes down.