Implemented Follow state
This is a very versatile state that allows you to instruct an NPC to follow a given transform at a given distance while configurably maintaining line of sight. However, it also allows you to do things like not move at all allowing for only looking at the transform. Or to exit when it gets within a radius to set up for things like the shoot interaction.