Added initial "hearing" implementation.
Currently I have no way of regulating the frequency at which "hearing" something interrupts the controller so something like footsteps causes NPCs to repeatedly comment on the fact that there are footsteps which is just weird. I am planning on making it configurable whether a certain sound causes repeated interrupts of the same type.