Added more filters to sound reactions
We can now filter by "sound type" which is currently only footsteps and "default". This allows NPCs to react differently to different types of fsounds.

We also have a bool that tells the sound handler whether it should repeatedly interrupt if a sound with the same reaction is heard multiple times while the interrupt graph is still running in order to prevent repeatedly reacting to things like footsteps.