Implemented main menu and scene loading
Also finished first implementation of ISaveable for NPCs and Interactables. NPCs currently do not save and load their state graphs or state controller. Cause those are hard and I want to know if the rest of the system works before I implement those.

To prep for the save load system I implemented a scene manager that currently has the ability to load a level. Next I will add in the ability to pass load a specific save file which includes the scene name as well as all the saveable data. I have already checked and the ResourceDataManager is fully populated at the end of the load scene function so it is all ready to populate the saved data. I just need to change the logic so that it unloads the current scene, moves to the loading screen, starts/restarts the scene, and then calls the load function with the deserialized data.