Save system working except StateGraphController
Ironed out the major bugs in loading. Mostly it came down to the NavMeshAgent state not getting updated properly on load in.

StateGraphController has not been implemented because the last time I tried to serialize nodes it broke everything.

Manager state has not been implemented. That should not be an issue.