Implemented player roles.
These are roles that, once any thrall has them, they are permanent. This is used to do things like implement "knowledge" that the parasite has. As a demo, I implemented the pet owner flow where walking near the pet owner causes you to gain knowledge about the PA system which allows you to interact with it to trigger the owner to pick up his food and leave. The knowledge is shared across all thralls so any thrall under your control is able to take the action.