Rewrote inventory UI to use new interactable menu
While doing this I also discovered that the inventory UI was horribly bloated and badly written. I think I just got an AI to do it.

I also created the "Shoot Player Action" custom action component to ensure that triggering custom actions works. It does. I didn't write the custom graph for the action though so for now it's the exact same as just the default.