Implemented simple manager for disembarking passengers and 2 example passengers
DisembarkingManager uses a PooledNpcManager to reset a group of NPCs that walk out of the ship the player starts at regular intervals. Right now it only resets them if every NPC is done with their routine, but once there are enough passenger NPCs I plan to just have it be at an interval and tune that interval so a good amount will be done at any given time. I might even make there be more than necessary and restrict the number that can be reset at any given time to act as a buffer.