Implemented wizards for easier creation of NPCs and NPC roles
Also started implementation of pet owner NPC. To do this faster I also implemented a new node that combines moving and using as those two are too linked to separate. This utility also uses control triggers to automatically get the move point instead of that having to be a separate blackboard variable.